# worker/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp Worker API",
    "description": "HTTP API for WhatsApp operations via whatsapp-web.js. Use either `Authorization: Bearer <API_TOKEN>` or the admin session cookie set by /auth/login.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "https://agente.imersaoparadigma.com.br" }],
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer" },
      "cookie": { "type": "apiKey", "in": "cookie", "name": "ww_session" }
    }
  },
  "security": [{ "bearer": [] }, { "cookie": [] }],
  "paths": {
...
    "/messages/private": {
      "post": {
        "summary": "Send text to private chat",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["to", "body"], "properties": { "to": { "type": "string" }, "body": { "type": "string" }, "typing": { "type": "boolean" }, "priority": { "type": "integer" } } } } } }
      }
    },
    "/messages/group": {
      "post": {
        "summary": "Send text to group",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["groupId", "body"], "properties": { "groupId": { "type": "string" }, "body": { "type": "string" } } } } } }
      }
    },
...
  }
}